Refactoring input parsers.
