+++ vires/vires/processes/util/parsers/filters.py
@@ -1,6 +1,6 @@
 #-------------------------------------------------------------------------------
 #
-#  Process Utilities
+# Process Utilities - filters input parsers
 #
 # Authors: Martin Paces <martin.paces@eox.at>
 #-------------------------------------------------------------------------------
@@ -24,16 +24,33 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #-------------------------------------------------------------------------------
+# pylint: disable=too-many-branches,unused-argument
 
-from .parsers import (
-    parse_style, parse_collections,
-    parse_model_expression, parse_model_list,
-    parse_filters,
-    parse_variables, get_subtracted_variables,
-)
-from .png_output import data_to_png, array_to_png
-from .resolver import VariableResolver, extract_product_names
-from .spacecraft_subtraction import group_subtracted_variables
-from .magnetic_model_renderer import (
-    get_extra_model_parameters, render_model, ALLOWED_VARIABLES,
-)
+import re
+from eoxserver.services.ows.wps.exceptions import InvalidInputValueError
+from ..filters import ScalarRangeFilter, VectorComponentRangeFilter
+
+
+RE_FILTER_NAME = re.compile(r'(^[^[]+)(?:\[([0-9])\])?$')
+
+
+def parse_filters(input_id, filter_string):
+    """ Parse filters' string and return list of the filter objects. """
+
+    def _get_filter(name, vmin, vmax):
+        match = RE_FILTER_NAME.match(name)
+        if match is None:
+            raise InvalidInputValueError(
+                input_id, "Invalid filter name %r" % name
+            )
+        variable, component = match.groups()
+        if component is None:
+            return ScalarRangeFilter(variable, vmin, vmax)
+        return VectorComponentRangeFilter(
+            variable, int(component), vmin, vmax
+        )
+
+    return [
+        _get_filter(name, vmin, vmax) for name, (vmin, vmax)
+        in parse_filters(input_id, filter_string).items()
+    ]
